pyproject.toml
[project]
name = "media-catalog-ms"
version = "0.1.0"
description = "Media Catalog Microservice"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.4",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.11.0",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.6",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-dotenv>=0.5.2",
    "polyfactory>=2.22.4",
    "ruff>=0.14.4",
    "httpx>=0.28.1",
]

[tool.pyright]
include=["app", "tests"]

[tool.ruff.lint]
per-file-ignores = { "__init__.py" = ["F401"] }
exclude = ["migrations/env.py", "tests/conftest.py"]
